:hammer: Move electron_driver fixture to conftest
